# entity-service/internal/handler/project_handler.go
// Copyright (c) 2026 WSO2 LLC. (https://www.wso2.com).
//
// WSO2 LLC. licenses this file to you under the Apache License,
// Version 2.0 (the "License"); you may not use this file except
// in compliance with the License.
// You may obtain a copy of the License at
//
// http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing,
// software distributed under the License is distributed on an
// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
// KIND, either express or implied.  See the License for the
// specific language governing permissions and limitations
// under the License.

// Package handler is declared in user_handler.go.
package handler

import (
	"encoding/json"
	"net/http"

	"github.com/wso2-open-operations/cs-tools/entity-service/internal/domain"
	"github.com/wso2-open-operations/cs-tools/entity-service/internal/service"
)

// ProjectHandler handles HTTP requests for the project resource.
type ProjectHandler struct {
	svc service.ProjectService
}

// NewProjectHandler constructs a ProjectHandler with the given service.
func NewProjectHandler(svc service.ProjectService) *ProjectHandler {
	return &ProjectHandler{svc: svc}
}

// SearchProjects handles POST /projects/search.
func (h *ProjectHandler) SearchProjects(w http.ResponseWriter, r *http.Request) {
	var req domain.SearchProjectsRequest
	if !decodeRequest(w, r, &req) {
		return
	}
	resp, err := h.svc.SearchProjects(r.Context(), req)
	if err != nil {
		writeServiceError(w, r, err)
		return
	}
	w.Header().Set("Content-Type", "application/json")
	_ = json.NewEncoder(w).Encode(resp)
}

// GetProject handles GET /projects/{id}.
func (h *ProjectHandler) GetProject(w http.ResponseWriter, r *http.Request) {
	id := r.PathValue("id")
	project, err := h.svc.GetProjectByID(r.Context(), id)
	if err != nil {
		writeServiceError(w, r, err)
		return
	}
	w.Header().Set("Content-Type", "application/json")
	_ = json.NewEncoder(w).Encode(project)
}

// ProjectUpdateHandler handles the project update endpoint, backed by the
// ServiceNow data source only.
type ProjectUpdateHandler struct {
	svc service.ProjectUpdateService
}

// NewProjectUpdateHandler constructs a ProjectUpdateHandler with the given service.
func NewProjectUpdateHandler(svc service.ProjectUpdateService) *ProjectUpdateHandler {
	return &ProjectUpdateHandler{svc: svc}
}

// UpdateProject handles PATCH /projects/{id}.
func (h *ProjectUpdateHandler) UpdateProject(w http.ResponseWriter, r *http.Request) {
	id := r.PathValue("id")
	var req domain.ProjectUpdateRequest
	if !decodeRequest(w, r, &req) {
		return
	}
	resp, err := h.svc.UpdateProject(r.Context(), id, req)
	if err != nil {
		writeServiceError(w, r, err)
		return
	}
	w.Header().Set("Content-Type", "application/json")
	_ = json.NewEncoder(w).Encode(resp)
}

// ProjectContactHandler handles HTTP requests for project contacts, backed by
// the ServiceNow data source only.
type ProjectContactHandler struct {
	svc service.ProjectContactService
}

// NewProjectContactHandler constructs a ProjectContactHandler with the given service.
func NewProjectContactHandler(svc service.ProjectContactService) *ProjectContactHandler {
	return &ProjectContactHandler{svc: svc}
}

// SearchProjectContacts handles POST /projects/{id}/contacts/search.
func (h *ProjectContactHandler) SearchProjectContacts(w http.ResponseWriter, r *http.Request) {
	id := r.PathValue("id")
	var req domain.SearchProjectContactsRequest
	if !decodeRequest(w, r, &req) {
		return
	}
	resp, err := h.svc.SearchProjectContacts(r.Context(), id, req)
	if err != nil {
		writeServiceError(w, r, err)
		return
	}
	w.Header().Set("Content-Type", "application/json")
	_ = json.NewEncoder(w).Encode(resp)
}

// GetProjectContact handles GET /projects/{id}/contacts/{contactId}.
func (h *ProjectContactHandler) GetProjectContact(w http.ResponseWriter, r *http.Request) {
	resp, err := h.svc.GetProjectContact(r.Context(), r.PathValue("id"), r.PathValue("contactId"))
	if err != nil {
		writeServiceError(w, r, err)
		return
	}
	w.Header().Set("Content-Type", "application/json")
	_ = json.NewEncoder(w).Encode(resp)
}